requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pytest==8.3.5
python-dotenv==1.2.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
